+++ routes/order-route.ts
@@ -1,65 +1,23 @@
-const express = require('express');
+import * as express from 'express';
 const router = express.Router();
-const { checkSchema } = require('express-validator');
-const { order } = require('../services');
-const { util } = require('../shared/lib');
-const logger = require('../shared/logger');
-const paymentSchema = require('./validation-schema');
+import util from '../shared/lib/index';
+import { checkSchema, ValidationChain } from 'express-validator';
+import { paymentSchema } from './validation-schema';
+import orderController from '../controller/order-controller';
 
 /* 주문 목록 조회 */
 router.get('/',
     [
         util.verifyToken,
-    ],
-    async (req, res) => {
-        let { userId } = req.user;
-        logger.reportRequest(req.url, req.method);
-        try {
-            let result = {};
-            result.data = await order.getOrderList(userId);
-            result.message = 'Success';
-            logger.reportResponse(req.url, req.method, result);
-            res.status(200).json(result);
-        } catch (e) {
-            logger.reportReponseErr(req.url, req.method, e);
-        }
-    })
+    ], orderController.getOrderList)
     .post('/', /* 결제 하기 */
         [
             checkSchema(paymentSchema),
             util.validate,
             util.verifyToken,
-        ],
-        async (req, res) => {
-            logger.reportRequest(req.url, req.method);
-            let { userId } = req.user;
-
-            try {
-                let result = {};
-                await order.handlePayment(userId, req.body);
-                result.message = 'Success';
-                logger.reportResponse(req.url, req.method, result);
-                res.status(200).json(result);
-            } catch (e) {
-                logger.reportReponseErr(req.url, req.method, e);
-            }
-        })
+        ], orderController.orderPayment)
     .get('/:orderId', /* 주문 상세 조회 */
         [
             util.verifyToken,
-        ],
-        async (req, res) => {
-            logger.reportRequest(req.url, req.method);
-            let { userId } = req.user;
-            let { orderId } = req.params;
-            try {
-                let result = {};
-                result.data = await order.getOrderDetail(userId, orderId);
-                result.message = 'Success';
-                logger.reportResponse(req.url, req.method, result);
-                res.status(200).json(result);
-            } catch (e) {
-                logger.reportReponseErr(req.url, req.method, e);
-            }
-        })
+        ], orderController.getOrderDetail)
 module.exports = router;
