+++ routes/user-route.ts
@@ -1,10 +1,8 @@
-const express = require('express');
+import express from 'express';
 const router = express.Router();
-const { body } = require('express-validator');
-const { users } = require('../services');
-const { util } = require('../shared/lib');
-const logger = require('../shared/logger');
-const jwt = require('jsonwebtoken');
+import { body } from 'express-validator';
+import util from '../shared/lib';
+import userController from '../controller/user-controller';
 
 router.post('/join',
     [
@@ -14,19 +12,8 @@ router.post('/join',
         body('password').notEmpty().withMessage('No password'),
         body('name').notEmpty().withMessage('No name'),
         util.validate
-    ],
-    async (req, res, next) => {
-        logger.reportRequest(req.url, req.method);
-        const userInfo = req.body;
-        try {
-            let result = await users.join(userInfo);
-            logger.reportResponse(req.url, req.method, result)
-            res.status(200).json(result);
-        } catch (e) {
-            logger.reportReponseErr(req.url, req.method, e);
-            res.status(500).json({ message: 'Server error' });
-        }
-    })
+    ], userController.join)
+
     .post('/login',
         [
             body('email')
@@ -34,34 +21,7 @@ router.post('/join',
                 .isEmail().withMessage(`Wrong email`),
             body('password').notEmpty().withMessage('No password'),
             util.validate
-        ],
-        async (req, res, next) => {
-            logger.reportRequest(req.url, req.method);
-            const loginInfo = req.body;
-
-            try {
-                let result = await users.login(loginInfo);
-                if (result.message === 'Success') {
-                    const token = jwt.sign({
-                        userId: result.userId,
-                        email: loginInfo.email,
-                        name: loginInfo.name
-                    }, process.env.PRIVATE_KEY, {
-                        expiresIn: '30m',
-                        issuer: "Anna"
-                    });
-                    logger.reportResponse(req.url, req.method, result)
-                    res.cookie("token", token);
-                    res.status(200).json(result);
-                }
-                else
-                    res.status(400).json(result);
-            } catch (e) {
-                logger.reportReponseErr(req.url, req.method, e);
-                res.status(500).json({ message: 'Server error' });
-            }
-        }
-    )
+        ], userController.login)
 
 router.post('/reset',
     [
@@ -69,47 +29,12 @@ router.post('/reset',
             .notEmpty().withMessage('No email')
             .isEmail().withMessage(`Wrong email`),
         util.validate
-    ],
-    async (req, res, next) => {
-        logger.reportRequest(req.url, req.method);
-        const { email } = req.body;
-        try {
-            let result = await users.isEmailMatch(email);
-            if (result.message === 'Success') {
-                const token = jwt.sign({
-                    email: email,
-                }, process.env.PRIVATE_KEY, {
-                    expiresIn: '30m',
-                    issuer: "Anna"
-                });
-                res.cookie("token", token);
-                res.status(200).json(result);
-            }
-            logger.reportResponse(req.url, req.method, result)
-        } catch (e) {
-            logger.reportReponseErr(req.url, req.method, e);
-            res.status(500).json({ message: 'Server error' });
-        }
-    })
+    ], userController.matchEmailForReset)
     .put('/reset',
         [
             body('password').notEmpty().withMessage('No password'),
             util.validate,
             util.verifyToken,
-        ],
-        async (req, res, next) => {
-            logger.reportRequest(req.url, req.method);
-            const { password } = req.body;
-            const { userId } = req.user;
-            try {
-                let result = await users.updatePassword(userId, password);
-                logger.reportResponse(req.url, req.method, result)
-                res.status(200).json(result);
-            } catch (e) {
-                logger.reportReponseErr(req.url, req.method, e);
-                res.status(500).json({ message: 'Server error' });
-            }
-        }
-    )
+        ], userController.reset)
 
-module.exports = router;
+export default router;
