--- shared/lib/index.js
@@ -1,5 +1,0 @@
-const util = require('./util');
-
-module.exports = {
-    util,
-};
