+++ shared/lib/index.ts
@@ -1,0 +1,3 @@
+import util from './util';
+
+export default util;
