+++ shared/lib/util.ts
@@ -1,14 +1,14 @@
-const { validationResult } = require('express-validator');
-const logger = require('../logger');
-const jwt = require('jsonwebtoken');
+import { validationResult } from 'express-validator';
+import logger from '../logger';
+import jwt from 'jsonwebtoken';
 
-module.exports = {
+export default {
     validate: function validate(req, res, next) {
         const route = req.originalUrl;
         const err = validationResult(req);
 
         if (!err.isEmpty()) {
-            logger.reportReponseErr(req.url, req.method, err);
+            logger.reportResponseErr(req.url, req.method, err);
             return res.status(401).json(err.array())
         } else {
             return next();
@@ -36,7 +36,7 @@ module.exports = {
                 return res.status(403).json(result);
             } else {
                 result = { message: 'Server Error' };
-                logger.reportReponseErr(req.url, req.method, e);
+                logger.reportResponseErr(req.url, req.method, e);
                 return res.status(500).json({ message: 'Server Error' });
             }
         }
